requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.0.0
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.3
mysqlclient==2.1.1
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
